FIX - Validasi user

+++ src/store/user.js
@@ -25,8 +25,10 @@ export const useUserStore = defineStore("user", {
           );
         })
         .catch((error) => {
-          window.location.reload();
-          localStorage.clear();
+          if (error.response.status === 401) {
+            localStorage.clear();
+            window.location.reload();
+          }
         });
     },
   },
